Updatde raw resources OK. TODO: Navigation on OpenDialog
